audio_processing/requirements.txt
fastapi
uvicorn
numpy
torch
whisper
pyannote.audio
python-multipart
websockets